+++ public/scripts/classes/game.js
